Initial Commit
Currently works, unsure if this is the correct way to do this though.

At the moment the router file is in the api folder, should probably be moved to its own folder or the root.

Should probably figure out a better way to determine different endpoints in a single file, or if it is better to make individual files for each endpoint.